# blink_swagger_docs.json
{
	"swagger": "2.0",
	"info": {
		"version": "1.4",
		"title": "Blink API",
		"contact": {}
	},
	"host": "blink-backend.herokuapp.com",
	"basePath": "/api",
	"securityDefinitions": {},
	"schemes": ["https"],
	"consumes": ["application/json"],
	"produces": ["application/json"],
	"paths": {
...
		"/auth/signout": {
			"get": {
				"summary": "signout",
				"tags": ["Authentication"],
				"operationId": "signout",
				"deprecated": false,
				"produces": ["application/json"],
				"parameters": [
					{
						"name": "Authorization",
						"in": "header",
						"required": true,
						"default": "Bearer {token}",
						"type": "string"
					}
				],
				"responses": {
					"200": {
						"description": "",
						"headers": {}
					}
				}
			}
		},
...
	},
...
}